Extract function to filter instance secrets by role
